+++ common/rmi/src/main/java/is/codion/common/rmi/server/DefaultServerAdmin.java
@@ -18,7 +18,6 @@
  */
 package is.codion.common.rmi.server;
 
-import is.codion.common.Separators;
 import is.codion.common.property.PropertyStore;
 import is.codion.common.property.PropertyStore.PropertyFormatter;
 import is.codion.common.rmi.client.ConnectionRequest;
@@ -32,6 +31,7 @@ import javax.management.Notification;
 import javax.management.NotificationEmitter;
 import javax.management.NotificationListener;
 import javax.management.openmbean.CompositeData;
+import java.io.File;
 import java.io.Serial;
 import java.io.Serializable;
 import java.lang.management.GarbageCollectorMXBean;
@@ -382,7 +382,7 @@ public class DefaultServerAdmin extends UnicastRemoteObject implements ServerAdm
 		@Override
 		public String formatValue(String property, String value) {
 			if (classOrModulePath(property) && !value.isEmpty()) {
-				return "\n" + String.join("\n", value.split(Separators.PATH_SEPARATOR));
+				return "\n" + String.join("\n", value.split(File.pathSeparator));
 			}
 
 			return value;