+++ changelog.md
@@ -4,6 +4,7 @@ Codion Change Log
 ## 0.18.9
 ### is.codion.common.core
 - Memory utility class removed.
+- Separators utility class removed.
 ### is.codion.common.rmi
 - ServerConfiguration.Builder.connectionMaintenanceIntervalMs() renamed to connectionMaintenanceInterval().
 - ServerAdmin, ServerStatistics.allocatedMemory() renamed totalMemory().